fixed terminology in first order: 	literal - input/end-point/cell => Xi 	term    - subsequence enclosed with xor (print-ascii [inside]) => X1X2X3...Xi XOR ... 	formula - one expression => y0 = X1X2... XOR ... XOR ...
code was also modified to support this terminology, parameter `literals` is now `terms` etc.

new:
	added arity, determines how many inputs is in one term so basicaly arity = 2 is how many gates can be active => 0,1,0,-1 is valid but 1,1,1,0 is not (1 or -1 occures > 2)
	added some new truth tables, comparator, aheadadder, median

+++ src/utils.hpp
@@ -29,7 +29,7 @@ namespace utils {
     template<class T>
     inline T &get_fittest_invidiual(std::vector<T> &population) {
         T *best = &population.at(0);
-        for (unsigned int i = 1; i < population.size(); i++) {
+        for (size_t i = 1; i < population.size(); i++) {
             if (population[i].fitness <= best->fitness) {
                 best = &population[i];
             }
